twinkle: Update vector classes (vectorMenu -> vector-menu) for the Twinkle menu
https://github.com/Amorymeltzer/twinkle/commit/d934c1fa25cb79b6f3716d491e897f2761ef22c0

The work is a little spread out, but see https://phabricator.wikimedia.org/T253329, https://phabricator.wikimedia.org/T254798, and https://phabricator.wikimedia.org/T254797.

Basically, for our purposes:

- `vectorMenu` replaced with `vector-menu`
    - Dropdown also gets `vector-menu-dropdown`
        - twinkle.css updated to be specific so that tabs doesn't break (it is now?)
    - Tabs also gets `vector-menu-tabs`
- `vectorMenuCheckbox` replaced with `vector-menu-checkbox`
- Add an innerDiv with `vector-menu-content`
- `menu` replaced with `vector-menu-content-list`, although this isn't unique

A lot of this is in flux at the moment (e.g. `emptyPortlet` -> `vector-menu-empty`).  The Vector skin is keeping all the old classes for now, but I'd like not to create (even) more work.